requirements.txt
# Core Flask Dependencies
Flask==3.1.2
flask-cors==6.0.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.3.0
blinker==1.9.0
itsdangerous>=2.2.0
colorama==0.4.6

# FastAPI Service Dependencies
fastapi>=0.110.0
uvicorn>=0.30.0

# LangChain and AI Dependencies
langchain>=0.2.11
langchain-core>=0.2.16
langchain-community>=0.2.0
langchain-groq>=0.1.5
langchain-huggingface>=0.0.3
langgraph>=0.1.5
langchain-text-splitters==0.3.2

# Data Processing and Analysis
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
typing-extensions>=4.5.0

# Vector Database and Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# PDF Processing
pypdf>=3.0.0
pdfplumber>=0.9.0

# Environment and Configuration
python-dotenv==1.0.0

# Text Processing
regex==2024.9.11

# Development and Testing (Optional)
jupyter>=1.0.0
pytest>=7.0.0

# Additional utilities
pathlib2>=2.3.7
